Upstream: gdhcp: Set client state to INIT_SELECTING on restart
With the previous behavior, we would end up sending a discover message
and then ignoring the reply due to the client state mismatch, so a
restart would never result in successful acquisition of a lease.

RFC 2131 specifies that the client should return to INIT state in
various restart scenarios.  However, also note that it would not make
sense in any case to send a discover message and ignore the reply.

Change-Id: Ia88452fb2d5e9670c81c58566d98f6fc21d9b198